refactor: rename layout_type_config to config
