+++ src/ingestion_pipeline/chunking/document_chunker_factory.py
@@ -3,8 +3,8 @@
 import logging
 
 from ingestion_pipeline.chunking.chunk_strategy import ChunkStrategy
+from ingestion_pipeline.chunking.strategies.layout.config import LayoutChunkingConfig
 from ingestion_pipeline.chunking.strategies.layout.layout_chunk_handler import TextractLayoutDocumentChunker
-from ingestion_pipeline.chunking.strategies.layout.layout_chunking_config import LayoutChunkingConfig
 from ingestion_pipeline.chunking.strategies.layout.types.key_value.layout_key_value import KeyValueChunker
 from ingestion_pipeline.chunking.strategies.layout.types.layout_text import LayoutTextChunkingStrategy
 from ingestion_pipeline.chunking.strategies.layout.types.list.list_chunker import LayoutListChunkingStrategy
