+++ src/ingestion_pipeline/chunking/strategies/layout/types/table/base.py
@@ -8,7 +8,7 @@ from textractor.entities.bbox import BoundingBox
 from textractor.entities.layout import Layout
 
 from ingestion_pipeline.chunking.schemas import DocumentChunk, DocumentMetadata
-from ingestion_pipeline.chunking.strategies.layout.layout_chunking_config import LayoutChunkingConfig
+from ingestion_pipeline.chunking.strategies.layout.config import LayoutChunkingConfig
 
 logger = logging.getLogger(__name__)
