+++ src/ingestion_pipeline/chunking/strategies/layout/types/table/layout_table.py
@@ -8,7 +8,7 @@ from textractor.entities.table import Table
 
 from ingestion_pipeline.chunking.exceptions import ChunkException
 from ingestion_pipeline.chunking.schemas import DocumentChunk, DocumentMetadata
-from ingestion_pipeline.chunking.strategies.layout.layout_chunking_config import LayoutChunkingConfig
+from ingestion_pipeline.chunking.strategies.layout.config import LayoutChunkingConfig
 from ingestion_pipeline.chunking.strategies.layout.types.base import LayoutType
 from ingestion_pipeline.chunking.strategies.layout.types.table.base import BaseTableChunker
