+++ tests/chunking/strategies/table/test_layout_table.py
@@ -7,7 +7,7 @@ from textractor.entities.layout import Layout, Line
 from textractor.entities.table import Table
 
 from ingestion_pipeline.chunking.exceptions import ChunkException
-from ingestion_pipeline.chunking.strategies.layout.layout_chunking_config import LayoutChunkingConfig
+from ingestion_pipeline.chunking.strategies.layout.config import LayoutChunkingConfig
 from ingestion_pipeline.chunking.strategies.layout.types.table import LayoutTableChunkingStrategy
 from ingestion_pipeline.chunking.strategies.layout.types.table.cell_chunker import CellTableChunker
 from ingestion_pipeline.chunking.strategies.layout.types.table.line_chunker import LineTableChunker
