+++ tests/chunking/test_textract_layout_text_chunker.py
@@ -9,7 +9,7 @@ from textractor.entities.bbox import BoundingBox
 
 import ingestion_pipeline.chunking.strategies.layout.types.layout_text as layout_text_module
 from ingestion_pipeline.chunking.schemas import DocumentChunk, DocumentMetadata
-from ingestion_pipeline.chunking.strategies.layout.layout_chunking_config import LayoutChunkingConfig
+from ingestion_pipeline.chunking.strategies.layout.config import LayoutChunkingConfig
 from ingestion_pipeline.chunking.strategies.layout.types.layout_text import LayoutTextChunkingStrategy
 from ingestion_pipeline.chunking.utils.bbox_utils import combine_bounding_boxes
